requirements.txt
alembic
fastapi
psycopg2-binary
python-dotenv
sqlalchemy
fastapi-users[sqlalchemy]
fastapi[all]
uvicorn
asyncpg
fastapi-cache2[redis]
redis
gunicorn
celery
flower
pydantic
starlette
passlib.context
celery
redis